Fixing up tests
