+++ persist/filesystem/filesystem_bank_record_id_index_test.go
@@ -1,4 +1,4 @@
-package persist
+package filesystem
 
 import (
 	"io/ioutil"
