+++ persist/filesystem/filesystem_test.go
@@ -249,6 +249,8 @@ func TestFileSystem_ListBranches(t *testing.T) {
 	}
 }
 
+
+
 func BenchmarkFileSystem_RoundTrip(b *testing.B) {
 	benchDir := path.Join("testdata", "bench", "filesystem", "roundtrip")
 	repo, err := json.NewFileSystemRepository(benchDir)
