+++ persist/filesystem/testdata/test1/current.txt
