+++ persist/filesystem/testdata/test2/current.txt
