+++ persist/filesystem/testdata/test3/.baronial/current.txt
