+++ persist/filesystem/testdata/test3/.baronial/objects/1fd675c891e5bd9029475bc9e769a5566af02055.json
