+++ persist/filesystem/testdata/test3/.baronial/objects/434926f0aa5dde53c24c81e8061f99ab963bb98f.json
