+++ persist/filesystem/testdata/test3/.baronial/objects/4f2253b28451ee743b38b2bc83f696f39bb2a568.json
