+++ persist/filesystem/testdata/test3/.baronial/objects/c4b86366e5c86ff4877baedca728c8862479233f.json
