+++ persist/filesystem/testdata/test3/.baronial/objects/c76ece80f16ef8cfe0213ca4b568d7d576a96f9d.json
