+++ persist/filesystem/testdata/test3/.baronial/objects/d9ff31f511c1fce2ec75b5efc500a6dfb4d83452.json
