+++ persist/filesystem/testdata/test3/.baronial/objects/e159ed13c539eba96ddc805c5ccce223903f0427.json
