+++ persist/filesystem/testdata/test3/.baronial/objects/e21102183a9a693b2e04c6709c8161071dbe2936.json
