+++ persist/filesystem/testdata/test3/.baronial/objects/f3bd1757f2ccfe63eefce93ffe4c046166d0a1a5.json
