+++ persist/filesystem/testdata/test3/budget/cash.txt
