+++ persist/filesystem/testdata/test4/.baronial/current.txt
