+++ persist/filesystem/testdata/test4/.baronial/objects/07e72edcf913fd3ef5eababf60852216d68dbb90.json
