+++ persist/filesystem/testdata/test4/.baronial/objects/1fd675c891e5bd9029475bc9e769a5566af02055.json
