+++ persist/filesystem/testdata/test4/.baronial/objects/2bc73e4f09eaf508a9912615e5fc686a14712759.json
