+++ persist/filesystem/testdata/test4/.baronial/objects/32cae97c1e69c8365e54c9aeb2f551e4d134d052.json
