+++ persist/filesystem/testdata/test4/.baronial/objects/960a403e64cca0c8022c8d72e96905991b74f533.json
