+++ persist/filesystem/testdata/test4/.baronial/objects/dee71d098f58bf5005d5bf3844b6c23f63e92195.json
