+++ persist/filesystem/testdata/test4/accounts/checking/cash.txt
