+++ persist/filesystem/testdata/test4/budget/cash.txt
