+++ persist/json/loader_test.go
@@ -93,3 +93,41 @@ func TestLoadAncestor(t *testing.T) {
 		}
 	}
 }
+
+
+func TestCache_Load_reuseHits(t *testing.T) {
+	var err error
+	ctx := context.Background()
+	const rawTargetId = "07e72edcf913fd3ef5eababf60852216d68dbb90"
+	var targetId envelopes.ID
+	targetId.UnmarshalText([]byte(rawTargetId))
+
+	passThrough := filesystem.FileSystem{
+		Root: "../filesystem/testdata/test3/.baronial",
+	}
+
+	subject := persist.NewCache(10)
+	subject.Loader = &Loader{
+		Fetcher: passThrough,
+		Loopback: subject,
+	}
+
+	var want envelopes.State
+	err = subject.Load(ctx, targetId, &want)
+	if err != nil {
+		t.Error(err)
+		return
+	}
+
+	var got envelopes.State
+	err = subject.Load(ctx, targetId, &got)
+	if err != nil {
+		t.Error(err)
+		return
+	}
+
+	if got.Budget != want.Budget {
+		t.Logf("When encountering a cache hit, the SAME Budget object should be reused")
+		t.Fail()
+	}
+}
